server/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
bidict==0.22.1
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
cryptography==41.0.7
Django==5.0
django-request-logging==0.7.5
djangorestframework==3.14.0
dnspython==2.4.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
lxml==4.9.3
mypy-extensions==1.0.0
Naked==0.1.32
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
python-engineio==4.8.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
shellescape==3.8.1
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
unittest-xml-reporting==3.2.0
urllib3==2.1.0
wsproto==1.2.0
